Move backgroundClassName to utils/theme
